Address PR feedback: deduplicate fixture, keep ValueError
- Revert update_page_content to raise ValueError on missing page;
  update test to assert the raise instead of expecting silent no-op
- Move envelope_cleanup fixture to conftest.py, remove 3 duplicate
  copies (~75 lines removed)
